requirements.txt
click
docxtpl
structlog
tzlocal
natural
ruamel.yaml
